requirements.txt
psycopg2-binary==2.9.9
pgvector==0.2.5
numpy==2.0.1
pandas==2.2.2
rapidfuzz==3.9.6
streamlit==1.36.0
anthropic==0.34.2
python-dotenv==1.0.1
tqdm==4.66.4
